Testing with Jest and Puppeteer
This change adds:

* Jest as our test runner
* Puppeteer to drive our Chromium
* Helper methods to drive Puppeteer to introspect the DevTools panel UI
  generated by our Chrome extension
* A test server to serve a HTML page with Trusted Types headers, meta
  tags, and violations, all configurable through URL parameters
* Starter tests to test that the extension loads all of its UI properly
  and can be interacted with through Puppeteer

We are deprecating the previous changes:

* https://ise-web-review.git.corp.google.com/c/trusted-types-helper/+/1320
* https://ise-web-review.git.corp.google.com/c/trusted-types-helper/+/1280

in favor of this one. After evaluating Karma, Mocha, Protractor, and
Playwright, we settled on Jest with Puppeteer because this is the only
test framework that let us access the power user settings of Puppeteer
necessary to introspect the UIs generated by the extension. The other
testing frameworks prioritize using Puppeteer to run existing unit tests
in a real browser rather than using Puppeteer to run true integration
tests, which is what we want to do to follow Chrome DevRel's guidance on
testing extensions:
https://developer.chrome.com/docs/extensions/how-to/test/end-to-end-testing.

Change-Id: Ie0b18afe3a416b824247329b88986e46be555d89
Bug: 347982706